% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recur_dataset.R
\name{recur_dataset}
\alias{recur_dataset}
\title{function for dataset and graph
Recurpackage recur_dataset package
The input data set should be 4 columns,
The first three columns are the values of the three
numbers to be input in order to calculate the recursive sequence define by
X_n = X_(n-1) + |X_(n-2) -X(n-3)|/2  and the fourth column
is the positive integer n for the sequence to be generated.
The output is expected to be a line plot with X axis as n and Y axis as sequence numbers}
\usage{
recur_dataset(x)
}
\arguments{
\item{x}{4 columns wide dataset with the first three columns are the values of the three
numbers to be input}
}
\value{
return a line plot with x axis as the sequence length n and Y axis as the sequence number
vector output
}
\description{
function for dataset and graph
Recurpackage recur_dataset package
The input data set should be 4 columns,
The first three columns are the values of the three
numbers to be input in order to calculate the recursive sequence define by
X_n = X_(n-1) + |X_(n-2) -X(n-3)|/2  and the fourth column
is the positive integer n for the sequence to be generated.
The output is expected to be a line plot with X axis as n and Y axis as sequence numbers
}
\examples{
my_data <- tibble::tribble(
~x, ~y, ~z, ~n,
2,4,3,3,
2,4,3,5,
2,4,3,7,
2,4,3,10,
2,4,3,15,
2,4,3,20,
2,4,3,25
)

recur_dataset(my_data)
}
