% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recur_vec.R
\name{recur_vec}
\alias{recur_vec}
\title{function for vector recursive sequence calculation
Recurpackage recur_vec package
The input should have one vector and a sequence number indicator n ,
The vector should have more than one elements while the sequence number indicator n
should be a positive integar (>n)
The output is the final nth element of the sequence to calculate defined as following
X_n = X_(n-1) + |X_(n-2) -X(n-3)|/2}
\usage{
recur_vec(x, n)
}
\arguments{
\item{x}{x should be a vector with more than one elements. However, the function will only take
the first three elements of the vector as the initial value as the recursive sequence
initial value input}

\item{n}{n should be a positive integer number}
}
\value{
The output is the final nth element of the sequence to calculate defined as following
X_n = X_(n-1) + |X_(n-2) -X(n-3)|/2
}
\description{
function for vector recursive sequence calculation
Recurpackage recur_vec package
The input should have one vector and a sequence number indicator n ,
The vector should have more than one elements while the sequence number indicator n
should be a positive integar (>n)
The output is the final nth element of the sequence to calculate defined as following
X_n = X_(n-1) + |X_(n-2) -X(n-3)|/2
}
\examples{
recur_vec(x = c(2, 4, 3), n = 3)
recur_vec(x = c(2, 4, 3), n = 4)
recur_vec(x = c(2, 4, 3,5), n = 3)
}
