% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurpkg-package.R
\docType{package}
\name{recurpkg-package}
\alias{recurpkg}
\alias{recurpkg-package}
\title{recurpkg: Calculating a recursive sequence vector and possible dataset for plot}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Fangfei Liu \email{fangfeiliu0811@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
